ShopPanel Made, Factory for ShopPanel Instantiation Madee

+++ Assets/Game/Scripts/Shop/ShopItem.cs
@@ -5,6 +5,6 @@ using UnityEngine;
 public class ShopItem : ScriptableObject
 {
     [field: SerializeField] public Sprite Image { get; private set; }
-    [field: SerializeField, Range(0,100000)] public float Price { get; private set; }
+    [field: SerializeField, Range(0,100000)] public int Price { get; private set; }
 
 }
